aggregate usecase package
